Time to implements variants

+++ Runtime/Editing/Default/AddVoxelEdit.cs
@@ -10,7 +10,7 @@ public struct AddVoxelEdit : IVoxelEdit {
     [ReadOnly] public float3 center;
     [ReadOnly] public float strength;
     [ReadOnly] public float radius;
-    [ReadOnly] public ushort material;
+    [ReadOnly] public byte material;
     [ReadOnly] public bool writeMaterial;
 
     public JobHandle Apply(SparseVoxelDeltaData data) {
