+++ Runtime/Editing/Default/CuboidDynamicEdit.cs
@@ -8,7 +8,7 @@ using UnityEngine;
 public struct CuboidDynamicEdit : IDynamicEdit {
     [ReadOnly] public float3 center;
     [ReadOnly] public float3 halfExtents;
-    [ReadOnly] public ushort material;
+    [ReadOnly] public byte material;
     [ReadOnly] public bool writeMaterial;
     public bool Enabled => true;
