+++ Runtime/Editing/Default/CuboidVoxelEdit.cs
@@ -8,7 +8,7 @@ public struct CuboidVoxelEdit : IVoxelEdit {
     [ReadOnly] public float3 center;
     [ReadOnly] public float3 halfExtents;
     [ReadOnly] public float strength;
-    [ReadOnly] public ushort material;
+    [ReadOnly] public byte material;
     [ReadOnly] public bool writeMaterial;
 
     public JobHandle Apply(SparseVoxelDeltaData data) {
